# Compute shrink-to-fit width for flex blocks