--- a/Source/Core/Layout/ContainerBox.cpp
+++ b/Source/Core/Layout/ContainerBox.cpp
@@ -31,6 +31,7 @@
 #include "../../../Include/RmlUi/Core/Element.h"
 #include "../../../Include/RmlUi/Core/ElementScroll.h"
 #include "../../../Include/RmlUi/Core/Profiling.h"
+#include "FlexFormattingContext.h"
 #include "FormattingContext.h"
 #include "LayoutDetails.h"
 #include <algorithm>
@@ -269,12 +270,12 @@ bool FlexContainer::Close(const Vector2f content_overflow_size, const Box& box, 
 
 float FlexContainer::GetShrinkToFitWidth() const
 {
-	// We don't currently support shrink-to-fit layout of flex containers. However, for the trivial case of a fixed
-	// width, we simply return that.
+	// For the trivial case of a fixed width, we simply return that.
 	if (element->GetComputedValues().width().type == Style::Width::Type::Length)
 		return box.GetSize().x;
 
-	return 0.0f;
+	// Infer shrink-to-fit width from the intrinsic width of the element.
+	return FlexFormattingContext::GetMaxContentSize(element).x;
 }
 
 String FlexContainer::DebugDumpTree(int depth) const