Refs https://github.com/mikke89/RmlUi/issues/552