@mikke89 2024-01-04
Thanks a lot for the pull request, enhancements to this part of the layout engine is very much welcome.

Can you expand a bit on how this relates to the specific issues in #552? Before and after comparisons would be helpful. If you are still working on this, let me know when you feel it is ready for review.