@mikke89 2024-01-21
Visual tests added here: ab20b36172ff8f9fd3df265bd89d1069ee3056e1

Also, added a benchmark: 37d0001d0923d6c37a36e32edb53a980d8ce4a3a

This PR didn't affect the numbers for our existing benchmarks. After adding some new shrink-to-fit cases, naturally we see a slow-down in these cases, compared to the before the changes in this PR, where we just didn't really format it.

Before:

| relative |               ns/op |                op/s |    err% |     total | Flexbox shrink-to-fit
|---------:|--------------------:|--------------------:|--------:|----------:|:----------------------
|   100.0% |            1,891.67 |          528,634.36 |    1.5% |      0.00 | `Reference`
|     8.4% |           22,620.00 |           44,208.66 |    1.3% |      0.00 | `Basic shrink-to-fit`
|     1.2% |          156,100.00 |            6,406.15 |    0.6% |      0.00 | `Nested shrink-to-fit`

After:

| relative |               ns/op |                op/s |    err% |     total | Flexbox shrink-to-fit
|---------:|--------------------:|--------------------:|--------:|----------:|:----------------------
|   100.0% |            1,894.64 |          527,803.96 |    0.4% |      0.00 | `Reference`
|     6.7% |           28,375.00 |           35,242.29 |    1.5% |      0.00 | `Basic shrink-to-fit`
|     0.1% |        2,911,200.00 |              343.50 |    0.3% |      0.03 | `Nested shrink-to-fit`

As discussed earlier, the nested case is definitely something to be aware of.